navigator: lint-enforce layer boundaries from ARCHITECTURE.md
Adds five no-restricted-imports rule blocks scoped per-directory to
formalize the layer dependency graph:

- maplibre-gl / react-map-gl/maplibre is import-able only from
  services/map/** (rule #3 in the doc).
- stores/** can't import controllers/services/views/components/
  reactions (rule #1).
- services/** can't import controllers/views/components (rule #2).
- reactions/** can't import controllers/views/components (rule #4).
- components/** can't import stores/controllers/services/reactions
  (rule #7) — extends the previous components/ rule.

allowTypeImports: true on each so prop-typing across layers stays OK.
ARCHITECTURE.md rule #6 (controller methods that are just `store.X = Y`
should move to the store) is not lintable; left as a code-review
concern.

Two pre-existing boundary leaks fixed in the same change:

- import 'maplibre-gl/dist/maplibre-gl.css' moved out of
  create-app.tsx into services/map/index.ts. The CSS belongs at the
  maplibre boundary, not at the composition root.
- lineColors + lineGradientExpression extracted from
  components/RoutesStyle.tsx to a new util/route-gradient.ts.
  RouteRenderer (in services/) was importing it from components/,
  which the new rule (rightly) flags. Both consumers — RoutesStyle
  and RouteRenderer — now import from util/.
