+++ eslint.config.mjs
@@ -69,6 +69,87 @@ export default [
       ],
     },
   },
+  // Navigator architecture: layer-boundary enforcement.
+  // See packages/apps/navigator/ARCHITECTURE.md for the rules.
+  {
+    files: ['packages/apps/navigator/src/**/*.{ts,tsx}'],
+    ignores: ['packages/apps/navigator/src/services/map/**'],
+    rules: {
+      'no-restricted-imports': [
+        'error',
+        {
+          patterns: [
+            {
+              group: ['maplibre-gl', 'react-map-gl/maplibre'],
+              message:
+                'services/map/ is the single maplibre boundary. Add a method to MapHandle/MapMarkers/MapCamera/MapStyle and call that.',
+              allowTypeImports: true,
+            },
+          ],
+        },
+      ],
+    },
+  },
+  {
+    files: ['packages/apps/navigator/src/stores/**/*.{ts,tsx}'],
+    rules: {
+      'no-restricted-imports': [
+        'error',
+        {
+          patterns: [
+            {
+              group: [
+                '**/controllers/**',
+                '**/services/**',
+                '**/views/**',
+                '**/components/**',
+                '**/reactions/**',
+              ],
+              message:
+                'stores/ are pure state + computeds. They must not depend on controllers/services/views/components/reactions.',
+              allowTypeImports: true,
+            },
+          ],
+        },
+      ],
+    },
+  },
+  {
+    files: ['packages/apps/navigator/src/services/**/*.{ts,tsx}'],
+    rules: {
+      'no-restricted-imports': [
+        'error',
+        {
+          patterns: [
+            {
+              group: ['**/controllers/**', '**/views/**', '**/components/**'],
+              message:
+                'services/ must not depend on controllers/views/components.',
+              allowTypeImports: true,
+            },
+          ],
+        },
+      ],
+    },
+  },
+  {
+    files: ['packages/apps/navigator/src/reactions/**/*.{ts,tsx}'],
+    rules: {
+      'no-restricted-imports': [
+        'error',
+        {
+          patterns: [
+            {
+              group: ['**/controllers/**', '**/views/**', '**/components/**'],
+              message:
+                'reactions/ wire stores → services. Controllers and views are not their concern.',
+              allowTypeImports: true,
+            },
+          ],
+        },
+      ],
+    },
+  },
   {
     files: ['packages/apps/navigator/src/components/**/*.{ts,tsx}'],
     rules: {
@@ -77,9 +158,14 @@ export default [
         {
           patterns: [
             {
-              group: ['**/stores/**', '**/controllers/**'],
+              group: [
+                '**/stores/**',
+                '**/controllers/**',
+                '**/services/**',
+                '**/reactions/**',
+              ],
               message:
-                'components/ are pure presenters. Move store/controller access to views/.',
+                'components/ are pure presenters. Take everything as props; move store/controller/service access to views/.',
               allowTypeImports: true,
             },
           ],
