+++ packages/apps/navigator/src/services/map/index.ts
@@ -1,3 +1,7 @@
+// Pull in maplibre's global CSS at the boundary, so consumers don't
+// need to remember to import it from the entry point.
+import 'maplibre-gl/dist/maplibre-gl.css';
+
 export { MapCamera } from './camera';
 export { MapHandle } from './handle';
 export { MapMarkers } from './markers';
